Reformatted and added some minor code for testing to the weatherCustom.py script
Normally I wouldn't do something as callous as reformat the code as it makes it harder to do diffs but this time it was a necessity. The format of the code is a mish-mash of indenting and I have a deep hatred of the python formatting. So I've reformatted the entire script. I've started adding the '#' indented to match things like if-else and try-except. I've also started adding code to allow local testing instead of pulling from the Weather Underground. I'm using a static json file for testing.